add lessons presentation layer, but no fix bug

+++ app/src/main/java/com/example/lessonslist/data/student/StudentItemDbModel.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class StudentItemDbModel(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
-	val paymentBalance: Float,
+	val paymentBalance: Int,
     val name: String,
     val lastname: String,
     val group: String,
